requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
fastapi==0.61.1
filterpy==1.4.5
gast==0.2.2
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.30.0
h11==0.11.0
h5py==2.10.0
idna==2.6
imageio==2.9.0
importlib-metadata==1.7.0
imutils==0.5.3
joblib==0.16.0
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
llvmlite==0.33.0
Markdown==3.2.2
matplotlib==3.3.0
mxnet-cu101==1.5.1.post0
mysql-connector==2.2.9
networkx==2.4
numba==0.50.1
numpy==1.19.1
opencv-python==4.2.0.32
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.12.4
pycrypto==2.6.1
pydantic==1.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.25
requests==2.24.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.5.2
SecretStorage==2.3.1
six==1.11.0
SQLAlchemy==1.3.19
starlette==0.13.6
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
timestring==1.6.4
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0